Remove independent article taxonomies from schemas
